🚧 TRN34K task: Fix merge gate hotspots
