+++ packages/agentplane/src/commands/context/release-readiness.test.ts
@@ -6,8 +6,6 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 
 import type { CommandContext } from "../shared/task-backend.js";
 import { cmdContextDoctor } from "./doctor.js";
-import { cmdContextExtractionApply } from "./extraction.js";
-import { cmdContextGraphValidate } from "./graph.js";
 import { cmdContextIngest } from "./ingest.js";
 import { cmdContextInit } from "./init.js";
 import { cmdContextReindex, readContextProjection } from "./reindex.js";
@@ -173,134 +171,6 @@ describe("context release readiness guards", () => {
     expect(first.title).toBe("context assimilation (explicit sources: first.md)");
     expect(second.title).toBe("context assimilation (explicit sources: second.md)");
     expect(first.title).not.toBe(second.title);
-  });
-
-  it("applies context_extraction SGR into non-empty derived facts and graph rows", async () => {
-    const root = await tempRoot();
-    await write(
-      root,
-      "context/extraction.json",
-      JSON.stringify(
-        {
-          schema_version: 1,
-          kind: "context_extraction",
-          task_id: "202605130501-CTXMAX",
-          reasoning: [
-            {
-              label: "entity-first",
-              summary: "Extract entities and relations before wiki synthesis.",
-            },
-          ],
-          source_refs: [{ path: "context/raw/research/source.md", lines: "1-6" }],
-          extracted_items: [
-            {
-              id: "entity.maximum-assimilation",
-              kind: "graph_entity",
-              summary: "Maximum assimilation is the canonical context workflow under test.",
-              source_refs: [{ path: "context/raw/research/source.md", lines: "1-2" }],
-              confidence: 0.9,
-              status: "accepted",
-              entity: {
-                id: "entity.maximum-assimilation",
-                kind: "concept",
-                label: "Maximum assimilation",
-                aliases: ["maximum-assimilation"],
-              },
-            },
-            {
-              id: "entity.derived-graph",
-              kind: "graph_entity",
-              summary: "Derived graph stores formal extraction rows.",
-              source_refs: [{ path: "context/raw/research/source.md", lines: "3-4" }],
-              confidence: 0.88,
-              status: "accepted",
-              entity: {
-                id: "entity.derived-graph",
-                kind: "concept",
-                label: "Derived graph",
-              },
-            },
-            {
-              id: "fact.maximum-assimilation.entity-first",
-              kind: "fact",
-              summary:
-                "Maximum assimilation requires formal entity and relation extraction before narrative wiki synthesis.",
-              source_refs: [{ path: "context/raw/research/source.md", lines: "1-6" }],
-              confidence: 0.92,
-              status: "accepted",
-            },
-            {
-              id: "edge.maximum-assimilation.produces.derived-graph",
-              kind: "graph_edge",
-              summary: "Maximum assimilation produces a derived graph layer.",
-              source_refs: [{ path: "context/raw/research/source.md", lines: "5-6" }],
-              confidence: 0.87,
-              status: "accepted",
-              edge: {
-                from: "entity.maximum-assimilation",
-                to: "entity.derived-graph",
-                relation: "produces",
-              },
-            },
-            {
-              id: "coverage.research-source",
-              kind: "coverage",
-              summary: "The source was covered by formal extraction rows before wiki synthesis.",
-              source_refs: [{ path: "context/raw/research/source.md", lines: "1-6" }],
-              confidence: 0.9,
-              status: "accepted",
-              coverage: {
-                source_path: "context/raw/research/source.md",
-                status: "covered",
-                reason: "Entities, fact, and relation rows cover the selected source.",
-                covered_item_ids: [
-                  "entity.maximum-assimilation",
-                  "entity.derived-graph",
-                  "fact.maximum-assimilation.entity-first",
-                  "edge.maximum-assimilation.produces.derived-graph",
-                ],
-              },
-            },
-          ],
-        },
-        null,
-        2,
-      ),
-    );
-
-    await cmdContextExtractionApply({
-      cwd: root,
-      parsed: { file: "context/extraction.json", taskId: "202605130501-CTXMAX", dryRun: false },
-    });
-    await cmdContextGraphValidate({ cwd: root, parsed: {} });
-
-    const facts = await readFile(
-      path.join(root, ".agentplane/context/derived/facts/facts.jsonl"),
-      "utf8",
-    );
-    const entities = await readFile(
-      path.join(root, ".agentplane/context/derived/graph/entities.jsonl"),
-      "utf8",
-    );
-    const edges = await readFile(
-      path.join(root, ".agentplane/context/derived/graph/edges.jsonl"),
-      "utf8",
-    );
-    const provenance = await readFile(
-      path.join(root, ".agentplane/context/derived/graph/provenance_edges.jsonl"),
-      "utf8",
-    );
-    const coverage = await readFile(
-      path.join(root, ".agentplane/context/derived/reports/coverage.jsonl"),
-      "utf8",
-    );
-
-    expect(facts).toContain("fact.maximum-assimilation.entity-first");
-    expect(entities).toContain("entity.maximum-assimilation");
-    expect(edges).toContain("edge.maximum-assimilation.produces.derived-graph");
-    expect(provenance).toContain("context/raw/research/source.md#lines=1-6");
-    expect(coverage).toContain("coverage.research-source");
-    expect(coverage).toContain("context/raw/research/source.md");
   });
 
   it("creates starter wiki structure on first context ingest with selected sources", async () => {
@@ -1110,61 +980,5 @@ describe("context release readiness guards", () => {
 
     const modulesIndex = await readFile(path.join(root, "context/wiki/modules/index.md"), "utf8");
     expect(modulesIndex).toContain("[Meridian Relay](meridian-relay.md)");
-  });
-
-  it("lints CRLF wiki frontmatter and rejects missing lint targets", async () => {
-    const root = await tempRoot();
-    await write(
-      root,
-      "context/wiki/decisions/crlf-page.md",
-      [
-        "---",
-        "agentplane_context:",
-        "  schema_version: 1",
-        "  artifact_type: wiki_page",
-        '  canonical_id: "wiki.decisions-crlf-page"',
-        '  title: "CRLF Page"',
-        "  modality: decision",
-        "  epistemic_status: sourced_claim",
-        "  source_refs: []",
-        "---",
-        "",
-        "# CRLF Page",
-        "",
-        "## Source References",
-        "",
-        "- no-source: local test fixture",
-        "",
-      ].join("\r\n"),
-    );
-    const out = vi.spyOn(process.stdout, "write").mockImplementation(() => true);
-
-    await cmdContextWikiLint({
-      cwd: root,
-      parsed: { path: "context/wiki/decisions/crlf-page.md" },
-    });
-    await expect(
-      cmdContextWikiLint({
-        cwd: root,
-        parsed: { path: "context/wiki/decisions/missing-page.md" },
-      }),
-    ).rejects.toThrow(/wiki lint target does not exist/u);
-
-    expect(out.mock.calls.map((call) => String(call[0])).join("")).toContain(
-      "context wiki lint: ok (1 page(s))",
-    );
-  });
-
-  it("requires AgentPlane frontmatter on initialized navigation files", async () => {
-    const root = await tempRoot();
-    await write(root, "context/wiki/AGENTS.md", "# Wiki Policy\n");
-    await write(root, "context/wiki/index.md", "# Project Wiki\n");
-
-    await expect(
-      cmdContextWikiLint({
-        cwd: root,
-        parsed: { path: "context/wiki" },
-      }),
-    ).rejects.toThrow(/missing YAML frontmatter/u);
   });
 });
